+++ public/directives/wz-table/wz-table-directive.js
@@ -13,7 +13,7 @@
 import template from './wz-table.html';
 import { uiModules } from 'ui/modules';
 import { DataFactory } from '../../services/data-factory';
-import { KeyEquivalenece } from '../../../util/csv-key-equivalence';
+import { KeyEquivalence } from '../../../util/csv-key-equivalence';
 import { calcTableRows } from './lib/rows';
 import { parseValue } from './lib/parse-value';
 import * as pagination from './lib/pagination';
@@ -35,7 +35,6 @@ app.directive('wzTable', function () {
       allowClick: '=allowClick',
       implicitFilter: '=implicitFilter',
       rowSizes: '=rowSizes',
-      extraLimit: '=extraLimit',
       emptyResults: '=emptyResults',
       customColumns: '=customColumns',
       implicitSort: '=implicitSort'
@@ -531,7 +530,7 @@ app.directive('wzTable', function () {
         const items = [];
         for (const key in item) {
           !excluded.includes(key) &&
-            items.push({ key: KeyEquivalenece[key] || key, value: item[key] });
+            items.push({ key: KeyEquivalence[key] || key, value: item[key] });
         }
         const props = {
           items,