+++ util/csv-key-equivalence.js
@@ -9,7 +9,7 @@
  *
  * Find more information about this on the LICENSE file.
  */
-export const KeyEquivalenece = {
+export const KeyEquivalence = {
   id: 'ID',
   timestamp: 'Timestamp',
   url: 'URL',
@@ -100,5 +100,7 @@ export const KeyEquivalenece = {
   rationale: 'Rationale',
   registry: 'Registry',
   date: 'Date',
-  value: 'Value'
+  value: 'Value',
+  location: 'Location',
+  mtu: 'MTU'
 };